Problem with a plus-sign in email address
Hi there,
the email address I have in my system conf has a plus-sign like "example+example@example.com". When I entered this address into the forgot manager password field, MODX responded with "Email not found" and no password mail was sent, because the plus-sign wasn't escaped in URL.
With the addition of "encodeURIComponent" in line 26 the problem seems to be solved. :-)